Add files via upload
The project revolves around scrapping data from web ,  it includes the download of HTML content, accessing the team names the opponent teams their scores respectively and the result per match of worldcup 2019. Further the project focuses on lining up the matches info in excel format and creating teams folders and subsequently creating the pdfs of the opponent played with the team.